Add dl extention to sparse framework and relocate generic code from split learning example
