+++ examples/split_learning/docker-compose.yml
@@ -7,6 +7,7 @@ services:
       WORKER_LISTEN_ADDRESS: '0.0.0.0'
     volumes:
       - .:/app
+      - ../../src:/usr/lib/sparse
 
   split_training_client:
     image: 'split_training:client.amd64'
@@ -14,3 +15,4 @@ services:
       MASTER_UPSTREAM_HOST: 'split_training_server'
     volumes:
       - .:/app
+      - ../../src:/usr/lib/sparse
